Debugger: Implement selective tree update in BehaviorTreeView
Flame graph showed performance bottleneck with recreating and redrawing the whole tree each frame.
This commit should improve debugger performance by a good margin.

+++ editor/debugger/behavior_tree_data.h
@@ -22,7 +22,7 @@ protected:
 
 public:
 	struct TaskData {
-		int id = 0;
+		uint64_t id = 0;
 		String name;
 		bool is_custom_name = false;
 		int num_children = 0;
@@ -31,7 +31,7 @@ public:
 		String type_name;
 		String script_path;
 
-		TaskData(int p_id, const String &p_name, bool p_is_custom_name, int p_num_children, int p_status, double p_elapsed_time, const String &p_type_name, const String &p_script_path) {
+		TaskData(uint64_t p_id, const String &p_name, bool p_is_custom_name, int p_num_children, int p_status, double p_elapsed_time, const String &p_type_name, const String &p_script_path) {
 			id = p_id;
 			name = p_name;
 			is_custom_name = p_is_custom_name;
